Много всяких изменений самой разной степени значимости 1. Исправил много разных багов 2. Добавил автоматическое исчезновение кнопок replay (кастинг) и play/pause (рейтинг) 3. Сделал возможность тихо перезагружать уведомления каждый раз, когда пользователь открывает этот экран (сейчас отключил) 4. Кэширование видео в личном профиле в 2 этапа 5. Добавил пример приветственного экрана с подсказками (сейчас отключен) и много мелких изменений

+++ AvatarAppIos/Models/System.swift
@@ -28,11 +28,14 @@ public class System {
             return
         }
         let fileManager = FileManager.default
-        do {
-            try fileManager.removeItem(at: url)
-        } catch {
-            print("Error deleting")
-            return
+        
+        if fileManager.fileExists(atPath: url.path) {
+            do {
+                try fileManager.removeItem(at: url)
+            } catch {
+                print("Error deleting")
+                return
+            }
         }
     }
